merge complete datacenter into main

+++ modules/project/projectHttpHandler/projectHttpHandler.go
@@ -8,6 +8,7 @@ import (
 	"RetroPGF-Hub/RetroPGF-Hub-Backend-Go/pkg/response"
 	"context"
 	"net/http"
+	"strconv"
 
 	"github.com/labstack/echo/v4"
 )
@@ -18,6 +19,7 @@ type (
 		FindOneProjectHttp(c echo.Context) error
 		UpdateOneProjectHttp(c echo.Context) error
 		DeleteOneProjectHttp(c echo.Context) error
+		FindAllProeject(c echo.Context) error
 	}
 
 	projectHttpHandler struct {
@@ -45,12 +47,12 @@ func (h *projectHttpHandler) CreateNewProjectHttp(c echo.Context) error {
 		return response.ErrResponse(c, http.StatusBadRequest, err.Error())
 	}
 	userId := c.Get("user_id").(string)
-	req.CreatedBy = userId
-	// email := c.Get("email")
-	// source := c.Get("source")
-	// fmt.Println(userId, email, source)
+	if len(userId) < 5 {
+		return response.ErrResponse(c, http.StatusBadRequest, "unauthorized user")
+	}
 
-	res, err := h.projectUsecase.CreateNewProjectUsecase(ctx, req)
+	req.CreatedBy = userId
+	res, err := h.projectUsecase.CreateNewProjectUsecase(ctx, &h.cfg.Grpc, req)
 	if err != nil {
 		return response.ErrResponse(c, http.StatusBadRequest, err.Error())
 	}
@@ -64,8 +66,10 @@ func (h *projectHttpHandler) CreateNewProjectHttp(c echo.Context) error {
 
 func (h *projectHttpHandler) FindOneProjectHttp(c echo.Context) error {
 	ctx := context.Background()
-
 	projectId := c.Param("projectId")
+	if len(projectId) < 5 {
+		return response.ErrResponse(c, http.StatusBadRequest, "projectId is required")
+	}
 
 	userId := c.Get("user_id").(string)
 
@@ -86,6 +90,9 @@ func (h *projectHttpHandler) DeleteOneProjectHttp(c echo.Context) error {
 	projectId := c.Param("projectId")
 
 	userId := c.Get("user_id").(string)
+	if len(userId) < 5 {
+		return response.ErrResponse(c, http.StatusBadRequest, "unauthorized user")
+	}
 
 	err := h.projectUsecase.DeleteOneProjectUsecase(ctx, projectId, userId)
 	if err != nil {
@@ -107,9 +114,16 @@ func (h *projectHttpHandler) UpdateOneProjectHttp(c echo.Context) error {
 	if err := wrapper.Bind(req); err != nil {
 		return response.ErrResponse(c, http.StatusBadRequest, err.Error())
 	}
+
 	projectId := c.Param("projectId")
+	if len(projectId) < 5 {
+		return response.ErrResponse(c, http.StatusBadRequest, "projectId is required")
+	}
 
 	userId := c.Get("user_id").(string)
+	if len(userId) < 5 {
+		return response.ErrResponse(c, http.StatusBadRequest, "unauthorized user")
+	}
 
 	res, err := h.projectUsecase.UpdateOneProjectUsecase(ctx, &h.cfg.Grpc, userId, projectId, req)
 	if err != nil {
@@ -119,5 +133,39 @@ func (h *projectHttpHandler) UpdateOneProjectHttp(c echo.Context) error {
 	return response.SuccessResponse(c, http.StatusOK, map[string]any{
 		"msg":     "ok",
 		"project": res,
+	})
+}
+
+func (h *projectHttpHandler) FindAllProeject(c echo.Context) error {
+	ctx := context.Background()
+	var limit, skip int = 20, 0
+
+	limitStr := c.QueryParam("limit")
+	skipStr := c.QueryParam("skip")
+	userId := c.Get("user_id").(string)
+	if limitStr != "" {
+		parsedLimit, err := strconv.Atoi(limitStr)
+		if err != nil {
+			return response.ErrResponse(c, http.StatusBadRequest, "limit incorrect format")
+		}
+		limit = parsedLimit
+	}
+
+	if skipStr != "" {
+		parsedSkip, err := strconv.Atoi(skipStr)
+		if err != nil {
+			return response.ErrResponse(c, http.StatusBadRequest, "skip incorrect format")
+		}
+		skip = parsedSkip
+	}
+
+	projects, err := h.projectUsecase.FindAllProjectDatacenterUsecase(ctx, &h.cfg.Grpc, int64(limit), int64(skip), userId)
+	if err != nil {
+		return response.ErrResponse(c, http.StatusBadRequest, err.Error())
+	}
+
+	return response.SuccessResponse(c, http.StatusOK, map[string]any{
+		"msg":     "ok",
+		"project": projects,
 	})
 }
